+++ Code/Engine/Foundation/Platform/Linux/OSFile_Linux.cpp
@@ -1,5 +1,6 @@
 #include <Foundation/FoundationPCH.h>
 
 #if EZ_ENABLED(EZ_PLATFORM_LINUX)
+#  include <Foundation/Platform/Posix/FileIterator_Posix.inl>
 #  include <Foundation/Platform/Posix/OSFile_Posix.inl>
 #endif
